Version 1.0 for release on PyPi
* Moved CLI into seam_erasure package
